* template commit

+++ src/game/game.cpp
@@ -1,4 +1,4 @@
 #include "game.h"
 
 Map *world = NULL;
-const Uint16 Game::PIXELS_PER_METER = 25;
+const Uint8 Game::PIXELS_PER_METER = 25;
